roachtest: create new kv/contention/nodes=4 test
Closes #32814.

This commit adds a roachtest that stresses the kinds of scenarios that
we saw bring down adriatic in #32582. It runs `kv` with a cycle length
of 512, a concurrency of 128, and a batch size of 4. The test runs this
on a 4 node cluster. It then verifies that QPS stayed above a certain
threshold for the entire duration of the test.

Release note: None
