+++ pkg/cmd/roachtest/cdc.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	gosql "database/sql"
 	"fmt"
-	"net/http"
 	"strconv"
 	"strings"
 	"sync/atomic"
@@ -28,8 +27,6 @@ import (
 	"github.com/Shopify/sarama"
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl"
 	"github.com/cockroachdb/cockroach/pkg/jobs/jobspb"
-	"github.com/cockroachdb/cockroach/pkg/ts/tspb"
-	"github.com/cockroachdb/cockroach/pkg/util/httputil"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/version"
@@ -884,75 +881,4 @@ func (c *topicConsumer) Close() {
 		}
 	}
 	_ = c.Consumer.Close()
-}
-
-func verifyTxnPerSecond(
-	ctx context.Context,
-	c *cluster,
-	t *test,
-	adminNode nodeListOption,
-	start, end time.Time,
-	txnTarget, maxPercentTimeUnderTarget float64,
-) {
-	// Query needed information over the timespan of the query.
-	adminURLs := c.ExternalAdminUIAddr(ctx, adminNode)
-	url := "http://" + adminURLs[0] + "/ts/query"
-	request := tspb.TimeSeriesQueryRequest{
-		StartNanos: start.UnixNano(),
-		EndNanos:   end.UnixNano(),
-		// Ask for one minute intervals. We can't just ask for the whole hour
-		// because the time series query system does not support downsampling
-		// offsets.
-		SampleNanos: (1 * time.Minute).Nanoseconds(),
-		Queries: []tspb.Query{
-			{
-				Name:             "cr.node.sql.txn.commit.count",
-				Downsampler:      tspb.TimeSeriesQueryAggregator_AVG.Enum(),
-				SourceAggregator: tspb.TimeSeriesQueryAggregator_SUM.Enum(),
-				Derivative:       tspb.TimeSeriesQueryDerivative_NON_NEGATIVE_DERIVATIVE.Enum(),
-			},
-			// Query *without* the derivative applied so we can get a total count of
-			// txns over the time period.
-			{
-				Name:             "cr.node.sql.txn.commit.count",
-				Downsampler:      tspb.TimeSeriesQueryAggregator_AVG.Enum(),
-				SourceAggregator: tspb.TimeSeriesQueryAggregator_SUM.Enum(),
-			},
-		},
-	}
-	var response tspb.TimeSeriesQueryResponse
-	if err := httputil.PostJSON(http.Client{}, url, &request, &response); err != nil {
-		t.Fatal(err)
-	}
-
-	// Drop the first two minutes of datapoints as a "ramp-up" period.
-	perMinute := response.Results[0].Datapoints[2:]
-	cumulative := response.Results[1].Datapoints[2:]
-
-	// Check average txns per second over the entire test was above the target.
-	totalTxns := cumulative[len(cumulative)-1].Value - cumulative[0].Value
-	avgTxnPerSec := totalTxns / float64(end.Sub(start)/time.Second)
-
-	if avgTxnPerSec < txnTarget {
-		t.Fatalf("average txns per second %f was under target %f", avgTxnPerSec, txnTarget)
-	} else {
-		t.l.Printf("average txns per second: %f", avgTxnPerSec)
-	}
-
-	// Verify that less than 5% of individual one minute periods were underneath
-	// the target.
-	minutesBelowTarget := 0.0
-	for _, dp := range perMinute {
-		if dp.Value < txnTarget {
-			minutesBelowTarget++
-		}
-	}
-	if perc := minutesBelowTarget / float64(len(perMinute)); perc > maxPercentTimeUnderTarget {
-		t.Fatalf(
-			"spent %f%% of time below target of %f txn/s, wanted no more than %f%%",
-			perc*100, txnTarget, maxPercentTimeUnderTarget*100,
-		)
-	} else {
-		t.l.Printf("spent %f%% of time below target of %f txn/s", perc*100, txnTarget)
-	}
 }
